Add one-way sync mode
Currently, the usage of --two-way option is obvious. User can use
it for both uploading and downloading notes. But there is no way
for downloading notes only, either a obvious or documented way for
uploading notes only. There may be users who want to only download
notes to local machine for backup purpose or search useful information
from notes more efficiently. This patch adds --one-way option for
that. It can be used for either uploading notes only or downloading
only.